fix:  Optimize style
1. searchbox
2. pagation
3. add personal info at siderbar
