+++ components/Sidebar.vue
@@ -1,5 +1,6 @@
 <template>
   <aside class="sidebar">
+    <PersonalInfo/>
     <NavLinks/>
     <slot name="top"/>
     <SidebarLinks :depth="0" :items="items"/>
@@ -9,12 +10,13 @@
 
 <script>
 import SidebarLinks from '@theme/components/SidebarLinks'
+import PersonalInfo from '@theme/components/PersonalInfo'
 import NavLinks from '@theme/components/NavLinks'
 
 export default {
   name: 'Sidebar',
 
-  components: { SidebarLinks, NavLinks },
+  components: { SidebarLinks, NavLinks, PersonalInfo },
 
   props: ['items']
 }
@@ -22,6 +24,8 @@ export default {
 
 <style lang="stylus">
 .sidebar
+  .personal-info-wrapper
+    display none
   ul
     padding 0
     margin 0
@@ -50,6 +54,8 @@ export default {
 
 @media (max-width: $MQMobile)
   .sidebar
+    .personal-info-wrapper
+      display block
     .nav-links
       display block
       .dropdown-wrapper .nav-dropdown .dropdown-item a.router-link-active::after
