+++ layouts/Tag.vue
@@ -24,7 +24,6 @@
       <!-- 分页 -->
       <ModuleTransition delay="0.16">
         <pagation
-          v-show="recoShowModule"
           class="pagation"
           :total="posts.length"
           :currentPage="currentPage"
