+++ layouts/Tags.vue
@@ -23,7 +23,6 @@
       <!-- 分页 -->
       <ModuleTransition delay="0.16">
         <pagation
-          v-show="recoShowModule"
           class="pagation"
           :total="$recoPosts.length"
           :currentPage="currentPage"
