+++ package.json
@@ -27,7 +27,7 @@
   "_from": "vuepress-theme-reco@0.2.1",
   "_resolved": "http://registry.npm.taobao.org/vuepress-theme-reco/download/vuepress-theme-reco-0.2.1.tgz",
   "dependencies": {
-    "@vuepress-reco/vuepress-plugin-back-to-top": "1.0.7",
+    "@vuepress-reco/vuepress-plugin-back-to-top": "1.0.8",
     "@vuepress-reco/vuepress-plugin-extract-code": "1.0.3",
     "@vuepress-reco/vuepress-plugin-loading-page": "1.0.5",
     "@vuepress-reco/vuepress-plugin-pagation": "1.0.6",
